+++ testsuite/forge-cli/src/main.rs
@@ -5,7 +5,7 @@ use anyhow::{format_err, Result};
 use aptos_logger::Level;
 use aptos_rest_client::Client as RestClient;
 use aptos_sdk::{move_types::account_address::AccountAddress, transaction_builder::aptos_stdlib};
-use forge::perf_success_criteria::SuccessCriteria;
+use forge::success_criteria::SuccessCriteria;
 use forge::{ForgeConfig, Options, *};
 use std::convert::TryInto;
 use std::sync::atomic::{AtomicBool, Ordering};
